requirements.txt
# Logging
python-json-logger==2.0.7

# Caching
django-redis==5.4.0
redis==5.0.1

# API Documentation
drf-spectacular==0.27.1

# Security
django-cors-headers==4.3.1
django-ratelimit==4.1.0

# API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Development
black==24.1.1
flake8==7.0.0
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0

# New additions
openai==1.12.0
setuptools>=68.0.0 